Remove `diff --git a/` from patches (#189)
The `diff --git a/` indicates that the patch is generated from `git diff`
and it should be applied with `patch -p1`. Because the patches bundled
with python-build have already re-formated for `patch -p0`, this is not
the desired behaviour.

Just removing `diff --git` from patches will force python-build to apply
those patches with `patch -p0`.