Run rehash command in the background
The rehash command doesn't take too long to run, but can add up when
several *env projects (rbenv, pyenv, nodenv, etc) are installed at the
same time. This makes it run in the background so that you don't have
to skip it altogether (using the `--no-rehash` flag) but also don't
have to wait a lot during shell initialization.

The `()` around the command prevent it from displaying the job number
or the "job done" message once it finishes.