Language tweaks to PULL_REQUEST_TEMPLATE.md
I saw one correction you had made and thought I could make this clearer as a native speaker.

"Prefer" is an interesting verb; I'm struggling to explain why "don't prefer" is not quite right. Saying you "don't prefer" something isn't the same as saying you "prefer not to do" something. Saying "don't prefer" is more like saying you don't have a preference, that is, you don't care if it happens or not. Saying "prefer not" is saying you have a preference not to do the thing, which is what I think you mean here.

Hope that is helpful, thanks for the useful project!