Fix malloc bin symlinks
Without this change, on FreeBSD, and potentially others, the binary gets build
with malloc and has a suffix of X.Ym when X.Y.0 is requested for install, for
example.  In this case, the verify_python script fails to find the pythonX.Y
binary and errors out of the installation, removing the successfully built
python version.  Here we include an additional symlink in the case where the
installed version contains a suffix that is not checked during the
verify_python call.