Pin bats to the latest release (1.2.0)
This fixes the following error when running `make test` with bats
installed from its master branch:

  /src/bats/libexec/bats-core/bats-exec-file:
  line 192:
  bats-exec-test:
  command not found

The Makefile currently runs bats from its master branch. This can lead
to errors when bats is broken between releases, as is currently the case
with bats at the following commit:

  bats-core/bats-core@b615ed8f750e45017b1ad070ef893d1e2552633a

Instead, use the latest release of bats, which is 1.2.0 (2020-04-25).