Remove build_package_ldflags_dirs()
build_package_ldflags_dirs() checks for the existence of directories and then attempts to create those directories in an area where it does not have permission. Per the comment by @chrahunt found [here](https://github.com/pyenv/pyenv/issues/732#issuecomment-464899679), this function was introduced to ruby-build for what is likely a Ruby-specific issue. I suggest removing this function entirely. If there is a good reason for keeping it, proper amendment should be made so it is functional for users utilizing LDFLAGS in their environment.