Support "arm64" identifier from brew coreutils' uname for Python 2.7/3.6+.
Python's `config.sub` attempts to determine whether the OS and machine
identifier type returned by `uname` is a valid target for compilation.
On M1 Macs, the system-default `uname` returns a value that's eventually
parsed as "arm-apple-darwin21.0.1", which passes the pattern matching.

However, if the coreutils Homebrew target is installed and its version
of `uname` is in the path, the result is "arm64-apple-darwin21.0.1".
This isn't handled by the existing patterns, and fails the build.

This appears to be a problem on all shipping versions of Python, tested
on Python 2.7.18, 3.6.15, 3.7.12, 3.8.12, 3.9.8, and 3.10.0 (the latest
versions in each of their respective series).

This change adds a patch to `config.sub` that adds `arm64-*` to the list
of patterns, alongside `arm-*`, allowing compilation to proceed. Patches
have been added for all the versions of Python mentioned above with the
exception of 3.6.15, which will need additional work in order to
compile.