Add --no-path-update option
Add option to skip PATH manipulation.  This is useful when one uses
pyenv --path in a login script, and wants to init pyenv functions
in each shell (i.e. zshrc).  However, in certain cases, like when
using pipenv, one needs zshrc to NOT push the pyenv shims to the
front of the path.